requirements.txt
plotly
openai
pandas
umap-learn
kaleido
kneed